requirements.txt
MouseInfo==0.1.3
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.3
pytweening==1.0.4
pywin32==303